+++ src/lib/core/port/primary/userpass-login-output-port.ts
@@ -2,7 +2,7 @@ import { UserpassLoginError, UserpassLoginResponse } from "../../data/userpass-l
 
 /**
  * Defines the output port for the userpass login use case. This interface muse be implemented by the
- * Presenter for the userpass login use case.
+ * Presenter for the userpass login use case. {@type T} is the type of the presenter i.e. HTTPResponse or other.
  */
 export default interface UserPassLoginOutputPort<T> {
     response: T;
